fix(dossier): sync search_terms after merge
Use a job instead of a lot of reloads
