+++ app/jobs/dossier_update_search_terms_job.rb
@@ -1,0 +1,8 @@
+class DossierUpdateSearchTermsJob < ApplicationJob
+  discard_on ActiveRecord::RecordNotFound
+
+  def perform(dossier)
+    dossier.update_search_terms
+    dossier.save!(touch: false)
+  end
+end
