+++ spec/models/concern/dossier_clone_concern_spec.rb
@@ -82,6 +82,10 @@ RSpec.describe DossierCloneConcern do
       it { expect { subject }.not_to change { dossier.reload.champs.order(:created_at).reject { _1.stable_id.in?([99, 994]) }.map(&:value) } }
       it { expect { subject }.to change { dossier.reload.champs.find { _1.stable_id == 99 }.value }.from('old value').to('new value') }
       it { expect { subject }.to change { dossier.reload.champs.find { _1.stable_id == 994 }.value }.from('old value').to('new value in repetition') }
+
+      it 'update dossier search terms' do
+        expect { subject }.to have_enqueued_job(DossierUpdateSearchTermsJob).with(dossier)
+      end
     end
 
     context 'with new revision' do
